Support deeply nested configurations
A configuration value can become deeply nested especially
when dealing with `NAPALM_ARGS`.  In this scenario the
approach to dump the configuration tree with `to_nice_json`
can lead to the creation of a configuration file that is not
valid Python anymore.  Store these values as strings in the
`configuration.py` file and load them with `json.loads`.