Use 'private' variable for netbox_config (#66)
This works around the problem in #66 that if you use the (increasingly more
common) include_role pattern variables you set externally could appear
immutable to the role logic. Thus, using set_fact with a combine() to
add SECRET_KEY to netbox_config will not work as expected.

Introducing a new 'private' (and thus mutable) variable in
load_variables works around the problem and allows the include_role
pattern to work.